Convert IDs to u32 and delays to f64

+++ src/graph_helper.rs
@@ -15,8 +15,8 @@ pub struct AudioGraphHelper {
 
     node_edges: FnvHashMap<NodeID, NodeEdges>,
 
-    next_node_id: u64,
-    next_edge_id: u64,
+    next_node_id: u32,
+    next_edge_id: u32,
 
     nodes_with_dirty_edges: FnvHashSet<NodeID>,
     needs_compile: bool,
@@ -53,7 +53,7 @@ impl AudioGraphHelper {
     /// Add a new [Node] the the audio graph.
     ///
     /// This will return the globally unique ID assigned to this node.
-    pub fn add_new_node(&mut self, latency: u64) -> NodeID {
+    pub fn add_new_node(&mut self, latency: f64) -> NodeID {
         let new_id = NodeID(self.next_node_id);
         self.next_node_id += 1;
 
@@ -84,7 +84,7 @@ impl AudioGraphHelper {
     ///
     /// This will return an error if a node with the given ID does not
     /// exist in the graph.
-    pub fn set_node_latency(&mut self, node_id: NodeID, latency: u64) -> Result<(), ()> {
+    pub fn set_node_latency(&mut self, node_id: NodeID, latency: f64) -> Result<(), ()> {
         let node = self.nodes.get_mut(&node_id).ok_or(())?;
 
         if node.latency != latency {
